code almost work

+++ internal/models/payment.go
@@ -1,14 +1,14 @@
 package models
 
 type Payment struct {
-	Transaction  string `json:"transaction"`
-	RequestID    string `json:"request_id"`
-	Currency     string `json:"currency"`
-	Provider     string `json:"provider"`
-	Amount       int    `json:"amount"`
-	PaymentDT    int    `json:"payment_dt"`
-	Bank         string `json:"bank"`
-	DeliveryCost int    `json:"delivery_cost"`
-	GoodsTotal   int    `json:"goods_total"`
-	CustomFee    int    `json:"custom_fee"`
+	Transaction  string  `json:"transaction"`
+	RequestID    string  `json:"request_id"`
+	Currency     string  `json:"currency"`
+	Provider     string  `json:"provider"`
+	Amount       float64 `json:"amount"`
+	PaymentDT    int     `json:"payment_dt"`
+	Bank         string  `json:"bank"`
+	DeliveryCost float64 `json:"delivery_cost"`
+	GoodsTotal   float64 `json:"goods_total"`
+	CustomFee    float64 `json:"custom_fee"`
 }
